Move modalMap and constants to utils folder
Updated imports after move.
