+++ src/components/modal/CurrentModal.tsx
@@ -1,8 +1,8 @@
 import { useSelector } from "react-redux";
-import { selectCurrentModal } from "~/store/selectors";
 
+import { selectCurrentModal } from "~/store/selectors";
 import ModalContainer from "~/components/modal/ModalContainer";
-import { MODAL_COMPONENTS } from "~/components/modal/modalMap";
+import { MODAL_COMPONENTS } from "~/utils/modalMap";
 
 const CurrentModal = () => {
   const currentModal = useSelector(selectCurrentModal);
@@ -10,6 +10,7 @@ const CurrentModal = () => {
   if (!currentModal) return null;
 
   const ModalComponent = MODAL_COMPONENTS[currentModal];
+
   return (
     <ModalContainer>
       <ModalComponent />
