--- src/types/constants.ts
@@ -1,9 +1,0 @@
-export enum ModalType {
-  AddBoard = "showAddBoardModal",
-  // EditBoard = "showEditBoardModal",
-  // DeleteBoard = "showDeleteBoardModal",
-  // ViewTask = "showViewTaskModal",
-  // AddTask = "showAddTaskModal",
-  // EditTask = "showEditTaskModal",
-  // DeleteTask = "showDeleteTaskModal",
-}
