+++ src/utils/constants.ts
@@ -1,0 +1,4 @@
+export enum ModalType {
+  AddBoard = "showAddBoardModal",
+  DeleteBoard = "showDeleteBoardModal",
+}
